backend/requirements.txt
Flask>=3.0
flask-cors>=4.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.6
psycopg2-binary>=2.9
pytest>=7.0
gunicorn>=21.0
